added happy path of query execution in snails db

+++ src/NlSqlBenchmark/snails/SnailsNlSqlBenchmark.py
@@ -1,7 +1,12 @@
-from NlSqlBenchmark.NlSqlBenchmark import NlSqlBenchmark
-import platform
 from os.path import dirname, abspath
+
 import docker
+import docker.models
+import docker.models.containers
+
+from NlSqlBenchmark.snails.util import db_util
+from NlSqlBenchmark.NlSqlBenchmark import NlSqlBenchmark
+
 
 
 class SnailsNlSqlBenchmark(NlSqlBenchmark):
@@ -10,29 +15,18 @@ class SnailsNlSqlBenchmark(NlSqlBenchmark):
     
     def __init__(
             self, 
-            db_host_profile="docker"
+            db_host_profile="docker",
+            kill_container_on_exit=False
             ):
         super().__init__()
         self.benchmark_folder = dirname(dirname(dirname(dirname(abspath(__file__))))) + "/benchmarks/snails"
+        self.kill_container_on_exit = kill_container_on_exit
+        self.container = None
         if db_host_profile == "docker":
-
-            try:
-                client = docker.from_env()
-                client.images.get('snails-db')
-            except docker.errors.ImageNotFound:
-                print("The 'snails-db' image is not available locally. Please run the install script in the benchmark folder before running the benchmark.")
-                raise
-
-            try:
-                print("Loading Docker container 'skaplel-running-snails'")
-                container = client.containers.get("skalpel-running-snails")
-                container.start()
-            except docker.errors.NotFound:
-                print("Container not found, running container for the first time on port 1433.")
-                client.containers.run("snails-db", ports={1433:1433}, name="skalpel-running-snails", detach=True)
-            self.container = container
-            db_info_file = self.benchmark_folder + "/ms_sql/dbinfo.json"
-
+            self.container = self._init_docker()
+            self.db_info_file = self.benchmark_folder + "/ms_sql/dbinfo.json"
+            
+        
         self.name = "SNAILS"
         self.naturalness = "Native"
         self.syntax = "tsql"
@@ -40,8 +34,9 @@ class SnailsNlSqlBenchmark(NlSqlBenchmark):
 
 
     def __del__(self):
-        print("SnailsNlSqlBenchmark is stopping Docker container 'skalpel-running-snails'")
-        self.container.kill()
+        if self.kill_container_on_exit:
+            print("SnailsNlSqlBenchmark is stopping Docker container 'skalpel-running-snails'")
+            self.container.kill()
 
 
     def __iter__(self):
@@ -60,6 +55,22 @@ class SnailsNlSqlBenchmark(NlSqlBenchmark):
         return 0
     
 
+    def _init_docker(self) -> docker.models.containers.Container:    
+        try:
+            client = docker.from_env()
+            client.images.get('snails-db')
+        except docker.errors.ImageNotFound:
+            print("The 'snails-db' Docker image is not available locally. Run the install script /benchmarks/snails/ms_sql/install_snails_db.sh before running the SNAILS benchmark.")
+            raise
+        try:
+            print("Loading Docker container 'skaplel-running-snails'")
+            container = client.containers.get("skalpel-running-snails")
+            container.start()
+        except docker.errors.NotFound:
+            print("Container not found, running container for the first time on port 1433.")
+            container = client.containers.run("snails-db", ports={1433:1433}, name="skalpel-running-snails", detach=True)
+        return container
+    
 
     def set_naturalness(self, naturalness: str) -> None:
         """
@@ -96,3 +107,18 @@ class SnailsNlSqlBenchmark(NlSqlBenchmark):
 
     def get_active_schema(self, database: str = None) -> dict:
         return super().get_active_schema(database)
+    
+
+
+    def execute_query(self, query: str, database: str = None, question: int = None) -> dict:
+        result_set_dict = db_util.do_query(
+            query=query,
+            database_name=database,
+            db_list_file=self.db_info_file
+        )
+        return {
+            "result_set": result_set_dict,
+            "database": database,
+            "question": question,
+            "error_message": ""
+        }
